Bug 703643: Make sure to handle overflow in pixel coordinate math.
The problem is that in jbig2_decode_refinement_template1_unopt() the
image width 3858760195 was read as an uint32_t, but is stored into
GRW as an int, which makes the value negative: -436207101.

Later on in the for loop x is set to 0, and then compared x < GRW,
which is true, causing jbig2_decode_refinement_template1_unopt() to
not decode any refinement region at all.

Later jbig2_page_add_result() and jbig2_image_compose() assume that
the refinement region was decoded properly and its pixels initialized,
causing use of uninitialized data when composing the refinement region
on top of the page.

This commit changes the types of GRW, GRH, x and y to be int64_t so
that they not only can contain the full range of image widths/heights,
but also may accepted the added GRREFERENCEDX/DY offsets. The
GRREFERENCEDX/DY offsets are originally parsed as int32_t, so this
commit also changes their variable types to int32_t.

As a consequence GRW in jbig2_decode_refinement_template1_unopt() for
the sample file now becoms 3858760195, which takes a long time to
process, but does not result in any MSAN complaint. If such large
images are not desirable, make sure to implement a memory limit
similar to jbig2dec -M.

This fixes OSS-fuzz 23762/42485273.

+++ jbig2dec/jbig2_generic.c
@@ -344,7 +344,7 @@ work similarly. */
 
 /* Get a bit. No bounds checking. */
 static inline int
-jbig2_image_get_pixel_fast(Jbig2Image *image, int x, int y)
+jbig2_image_get_pixel_fast(Jbig2Image *image, int64_t x, int64_t y)
 {
     const int byte = (x >> 3) + y * image->stride;
     const int bit = 7 - (x & 7);
@@ -367,10 +367,10 @@ jbig2_decode_generic_template0(Jbig2Ctx *ctx,
                                const Jbig2GenericRegionParams *params, Jbig2ArithState *as,
                                Jbig2Image *image, Jbig2ArithCx *GB_stats)
 {
-    const uint32_t GBW = image->width;
-    const uint32_t GBH = image->height;
+    const int64_t GBW = image->width;
+    const int64_t GBH = image->height;
     const uint32_t rowstride = image->stride;
-    uint32_t x, y;
+    int64_t x, y;
     byte *line2 = NULL;
     byte *line1 = NULL;
     byte *gbreg_line = (byte *) image->data;
@@ -435,10 +435,10 @@ jbig2_decode_generic_template0_unopt(Jbig2Ctx *ctx,
                                      Jbig2Segment *segment,
                                      const Jbig2GenericRegionParams *params, Jbig2ArithState *as, Jbig2Image *image, Jbig2ArithCx *GB_stats)
 {
-    const uint32_t GBW = image->width;
-    const uint32_t GBH = image->height;
+    const int64_t GBW = image->width;
+    const int64_t GBH = image->height;
     uint32_t CONTEXT;
-    uint32_t x, y;
+    int64_t x, y;
     int bit;
 
     if (pixel_outside_field(params->gbat[0], params->gbat[1]) ||
@@ -510,10 +510,10 @@ jbig2_decode_generic_template1_unopt(Jbig2Ctx *ctx,
                                      Jbig2Segment *segment,
                                      const Jbig2GenericRegionParams *params, Jbig2ArithState *as, Jbig2Image *image, Jbig2ArithCx *GB_stats)
 {
-    const uint32_t GBW = image->width;
-    const uint32_t GBH = image->height;
+    const int64_t GBW = image->width;
+    const int64_t GBH = image->height;
     uint32_t CONTEXT;
-    uint32_t x, y;
+    int64_t x, y;
     int bit;
 
     if (pixel_outside_field(params->gbat[0], params->gbat[1]))
@@ -578,10 +578,10 @@ jbig2_decode_generic_template1(Jbig2Ctx *ctx,
                                Jbig2Segment *segment,
                                const Jbig2GenericRegionParams *params, Jbig2ArithState *as, Jbig2Image *image, Jbig2ArithCx *GB_stats)
 {
-    const uint32_t GBW = image->width;
-    const uint32_t GBH = image->height;
+    const int64_t GBW = image->width;
+    const int64_t GBH = image->height;
     const uint32_t rowstride = image->stride;
-    uint32_t x, y;
+    int64_t x, y;
     byte *line2 = NULL;
     byte *line1 = NULL;
     byte *gbreg_line = (byte *) image->data;
@@ -643,10 +643,10 @@ jbig2_decode_generic_template2_unopt(Jbig2Ctx *ctx,
                                Jbig2Segment *segment,
                                const Jbig2GenericRegionParams *params, Jbig2ArithState *as, Jbig2Image *image, Jbig2ArithCx *GB_stats)
 {
-    const uint32_t GBW = image->width;
-    const uint32_t GBH = image->height;
+    const int64_t GBW = image->width;
+    const int64_t GBH = image->height;
     uint32_t CONTEXT;
-    uint32_t x, y;
+    int64_t x, y;
     int bit;
 
     if (pixel_outside_field(params->gbat[0], params->gbat[1]))
@@ -710,10 +710,10 @@ jbig2_decode_generic_template2(Jbig2Ctx *ctx,
                                 Jbig2Segment *segment,
                                 const Jbig2GenericRegionParams *params, Jbig2ArithState *as, Jbig2Image *image, Jbig2ArithCx *GB_stats)
 {
-    const uint32_t GBW = image->width;
-    const uint32_t GBH = image->height;
+    const int64_t GBW = image->width;
+    const int64_t GBH = image->height;
     const uint32_t rowstride = image->stride;
-    uint32_t x, y;
+    int64_t x, y;
     byte *line2 = NULL;
     byte *line1 = NULL;
     byte *gbreg_line = (byte *) image->data;
@@ -775,12 +775,12 @@ jbig2_decode_generic_template3(Jbig2Ctx *ctx,
                                Jbig2Segment *segment,
                                const Jbig2GenericRegionParams *params, Jbig2ArithState *as, Jbig2Image *image, Jbig2ArithCx *GB_stats)
 {
-    const uint32_t GBW = image->width;
-    const uint32_t GBH = image->height;
+    const int64_t GBW = image->width;
+    const int64_t GBH = image->height;
     const uint32_t rowstride = image->stride;
     byte *line1 = NULL;
     byte *gbreg_line = (byte *) image->data;
-    uint32_t x, y;
+    int64_t x, y;
 
 #ifdef OUTPUT_PBM
     printf("P4\n%d %d\n", GBW, GBH);
@@ -833,10 +833,10 @@ jbig2_decode_generic_template3_unopt(Jbig2Ctx *ctx,
                                      Jbig2Segment *segment,
                                      const Jbig2GenericRegionParams *params, Jbig2ArithState *as, Jbig2Image *image, Jbig2ArithCx *GB_stats)
 {
-    const uint32_t GBW = image->width;
-    const uint32_t GBH = image->height;
+    const int64_t GBW = image->width;
+    const int64_t GBH = image->height;
     uint32_t CONTEXT;
-    uint32_t x, y;
+    int64_t x, y;
     int bit;
 
     if (pixel_outside_field(params->gbat[0], params->gbat[1]))
@@ -901,10 +901,10 @@ jbig2_decode_generic_template0_TPGDON(Jbig2Ctx *ctx,
                                       Jbig2Segment *segment,
                                       const Jbig2GenericRegionParams *params, Jbig2ArithState *as, Jbig2Image *image, Jbig2ArithCx *GB_stats)
 {
-    const uint32_t GBW = image->width;
-    const uint32_t GBH = image->height;
+    const int64_t GBW = image->width;
+    const int64_t GBH = image->height;
     uint32_t CONTEXT;
-    uint32_t x, y;
+    int64_t x, y;
     int LTP = 0;
     int gmin, gmax;
     uint32_t left, right, top;
@@ -1101,10 +1101,10 @@ jbig2_decode_generic_template1_TPGDON(Jbig2Ctx *ctx,
                                       Jbig2Segment *segment,
                                       const Jbig2GenericRegionParams *params, Jbig2ArithState *as, Jbig2Image *image, Jbig2ArithCx *GB_stats)
 {
-    const uint32_t GBW = image->width;
-    const uint32_t GBH = image->height;
+    const int64_t GBW = image->width;
+    const int64_t GBH = image->height;
     uint32_t CONTEXT;
-    uint32_t x, y;
+    int64_t x, y;
     int LTP = 0;
 
     if (pixel_outside_field(params->gbat[0], params->gbat[1]))
@@ -1176,10 +1176,10 @@ jbig2_decode_generic_template2_TPGDON(Jbig2Ctx *ctx,
                                       Jbig2Segment *segment,
                                       const Jbig2GenericRegionParams *params, Jbig2ArithState *as, Jbig2Image *image, Jbig2ArithCx *GB_stats)
 {
-    const uint32_t GBW = image->width;
-    const uint32_t GBH = image->height;
+    const int64_t GBW = image->width;
+    const int64_t GBH = image->height;
     uint32_t CONTEXT;
-    uint32_t x, y;
+    int64_t x, y;
     int LTP = 0;
 
     if (pixel_outside_field(params->gbat[0], params->gbat[1]))
@@ -1251,10 +1251,10 @@ jbig2_decode_generic_template3_TPGDON(Jbig2Ctx *ctx,
                                       Jbig2Segment *segment,
                                       const Jbig2GenericRegionParams *params, Jbig2ArithState *as, Jbig2Image *image, Jbig2ArithCx *GB_stats)
 {
-    const uint32_t GBW = image->width;
-    const uint32_t GBH = image->height;
+    const int64_t GBW = image->width;
+    const int64_t GBH = image->height;
     uint32_t CONTEXT;
-    uint32_t x, y;
+    int64_t x, y;
     int LTP = 0;
 
     if (pixel_outside_field(params->gbat[0], params->gbat[1]))
